CIFAR10 reading

+++ nntest/nntest/data.h
@@ -13,30 +13,30 @@
 
 class Data {
 public:
-  Data(int size_dim0, int size_dim1) :
+  Data(long size_dim0, long size_dim1) :
     m_size_dim0(size_dim0),
     m_size_dim1(size_dim1) {};
   
   virtual ~Data() {};
   
-  int get_index(int idx_dim0, int idx_dim1) const {
-    int idx = idx_dim0*m_size_dim1+idx_dim1;
+  long get_index(long idx_dim0, long idx_dim1) const {
+    long idx = idx_dim0*m_size_dim1+idx_dim1;
     assert(idx < get_count());
     return idx;
   };
   
   virtual double* get_data() const = 0;
   
-  virtual double get_data_at(int row, int column) const {return get_data()[get_index(row, column)]; };
+  virtual double get_data_at(long row, long column) const {return get_data()[get_index(row, column)]; };
   
   virtual void copy_from(const Data& other) = 0;
   
   //return a data object with a reduced amount of rows
-  virtual Data* get_rows_slice(int start, int end) = 0;
+  virtual Data* get_rows_slice(long start, long end) = 0;
   
-  int get_count() const {return m_size_dim0*m_size_dim1;};
+  long get_count() const {return m_size_dim0*m_size_dim1;};
   
-  int get_size_dim(int dim) {
+  long get_size_dim(int dim) {
     switch (dim) {
       case 0:
         return m_size_dim0;
@@ -54,9 +54,9 @@ public:
   
 private:
   //size of the first dimension
-  int m_size_dim0;
+  long m_size_dim0;
   //size of the second dimension
-  int m_size_dim1;
+  long m_size_dim1;
 };
